Disable delete key, use relative paths in FLASK_BASE_URL, rename CSVNode to ItemsNode (#185)
* Disable delete key deleting a node in RF

* Change FLASK_BASE_URL to use relative path except in dev mode

* Rename CSVNode to ItemsNode and replace its icon.

* Update package version and rebuild react

* Add new Claude models
